Rename command module and update CLI banner
Fix incorrect import path and refresh CLI banner art. Updated openlinktoken_cli.commands to import OpenLinkTokenCommand from open_link_token_command (was open_token_command) and adjusted all tests to use the new module name. Also replaced the ASCII art banner in OpenLinkTokenCommand with an updated, more compact layout and ensured the version line remains displayed.
