+++ lib/python/openlinktoken-cli/src/main/openlinktoken_cli/commands/__init__.py
@@ -1,6 +1,6 @@
 # SPDX-License-Identifier: MIT
 
 # Import OpenLinkTokenCommand from its own module for better organization
-from openlinktoken_cli.commands.open_token_command import OpenLinkTokenCommand
+from openlinktoken_cli.commands.open_link_token_command import OpenLinkTokenCommand
 
 __all__ = ["OpenLinkTokenCommand"]
