+++ lib/python/openlinktoken-cli/src/test/openlinktoken_cli/test_initiate_exchange_command.py
@@ -15,7 +15,7 @@ import pytest
 
 from openlinktoken.exchange_jwe import decrypt_exchange_envelope
 from openlinktoken_cli.commands.initiate_exchange_command import InitiateExchangeCommand
-from openlinktoken_cli.commands.open_token_command import OpenLinkTokenCommand
+from openlinktoken_cli.commands.open_link_token_command import OpenLinkTokenCommand
 from openlinktoken_cli.util.ec_key_utils import SUPPORTED_CURVES, generate_key_pair, public_key_fingerprint
 from openlinktoken_cli.util.stdin_utils import read_required_env_bytes
