+++ lib/python/openlinktoken-cli/src/test/openlinktoken_cli/tokenize_command_demo_mode_test.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 
 import pytest
 
-from openlinktoken_cli.commands.open_token_command import OpenLinkTokenCommand
+from openlinktoken_cli.commands.open_link_token_command import OpenLinkTokenCommand
 from openlinktoken_cli.util.ec_key_utils import generate_key_pair
 
 # HMAC-SHA256 over 32 bytes → base64 → always exactly 44 characters
